"Crystal Abundance" fixes
Should be able to activate it even without any "Crystal Beast" monsters in your GY to Special Summon. Additionally, was not checking that there are other cards on the field to send to the GY (excluding what you would send as the cost). Lastly, should be able to chain the effect of "Ghost Belle & Haunted Mansion" to it, and made sure this new implementation works with "Destiny HERO - Diamond Dude".