Text updates for OCG functionality
The card texts of various cards have been updated to describe their functionality in the OCG (the scripts were already using OCG rulings). Cards that have received this update will have a note at the end of their card text box.

To see which cards were updated with this and for specific details on each card you can check the full list here:
https://yugioh-diffs.github.io/