Update core
0bb63fcf062172901f111ca347e3cc62d641d1c4
Check for the zone passed as second return value from the value function in effect_spsummon_procedure (Removes the need to do a redoundant check in the condition calling IsCanBeSpecialSummoned to explicitely pass the zone).
Add Card.GetAllCounters, Card.HasCounters and Card.RemoveAllCounters also remove the hardcoding for countertype = 0 in the other Counter functions.
Return the group itself when calling Group.Add and Group.Sub.