"Judge of the Ice Barrier" fix
Her effect shouldn't be applied because of the LP payment due to an effect like that of "Masquerade the Blazing Dragon". Also, fixed some cases where her effect was applied when cards without an LP payment cost were in the same Chain as a card with an LP payment cost.