"And the Band Played On" fix + associated cards updates/fixes
Simplified the script of "And the Band Played On" to fix some scenarios where Special Summons of monsters with the same Level at different timings was possible.

Updated/fixed the following cards that Special Summon multiple monsters sequentially but the summons are treated as happening at the same time:
- Cipher Spectrum
- Exosister Martha
- Mithra the Thunder Vassal
- Red-Eyes Retro Dragon
- Seed of Flame
- The Phantom Knights of Mist Claws

Should be able to use and fully resolve their effects even if the effect of "And the Band Played On" is applying + general fixes and script polishing all around.